Make it easier to call full submission validator from python
